Better support for snapshot versions
Introduces the ability to re-package snapshot versions into JAR files at
the location in which Maven would do so.

This is still incomplete and may not work properly. When testing with
PDFBox 4.0.0-SNAPSHOT, the plugin resolved it to 20240515 while
`mvn package` resolved it to 20240522, thus downloading that version and
using it without annotations. I'm not sure why exactly, so for the
time being, using Classport with snapshot versions may break.