Ignore test dependencies
This does not matter when producing dependency lists, as only the
included deps make it anyway. When reconstructing the tree however, if
dependency A has dependency T as test, but dependency B has T as
compile (so T makes it into the JAR), T will show up under A too.
Just filtering them out makes it so that A doesn't consider T a
dependency at all, thus avoiding the false positive.